+++ relay-server/src/services/processor/span/processing.rs
@@ -13,7 +13,7 @@ use crate::services::processor::{
     TransactionGroup, dynamic_sampling, event_type,
 };
 use crate::services::projects::project::ProjectInfo;
-use crate::utils::sample;
+use crate::utils;
 use chrono::{DateTime, Utc};
 use relay_base_schema::events::EventType;
 use relay_base_schema::project::ProjectId;
@@ -32,7 +32,7 @@ use relay_event_normalization::{
 };
 use relay_event_schema::processor::{ProcessingAction, ProcessingState, process_value};
 use relay_event_schema::protocol::{
-    BrowserContext, Event, EventId, IpAddr, Measurement, Measurements, Span, SpanData,
+    BrowserContext, CompatSpan, Event, EventId, IpAddr, Measurement, Measurements, Span, SpanData,
 };
 use relay_log::protocol::{Attachment, AttachmentType};
 use relay_metrics::{FractionUnit, MetricNamespace, MetricUnit, UnixTimestamp};
@@ -92,6 +92,7 @@ pub async fn process(
         geo_lookup,
     );
 
+    let org_id = managed_envelope.scoping().organization_id.value();
     let client_ip = managed_envelope.envelope().meta().client_addr();
     let filter_settings = &project_info.config.filter_settings;
     let sampling_decision = sampling_result.decision();
@@ -230,18 +231,40 @@ pub async fn process(
             }
         };
 
-        // Write back:
         let mut new_item = Item::new(ItemType::Span);
-        let payload = match annotated_span.to_json() {
-            Ok(payload) => payload,
-            Err(err) => {
-                relay_log::debug!("failed to serialize span: {}", err);
-                return ItemAction::Drop(Outcome::Invalid(DiscardReason::Internal));
-            }
-        };
-        new_item.set_payload(ContentType::Json, payload);
-        new_item.set_metrics_extracted(item.metrics_extracted());
+        if produce_compat_spans(&config, global_config, org_id) {
+            let span_v2 = annotated_span.map_value(relay_spans::span_v1_to_span_v2);
+            let compat_span = match span_v2.map_value(CompatSpan::try_from) {
+                Annotated(Some(Result::Err(err)), _) => {
+                    relay_log::debug!("failed to create compat span: {}", err);
+                    return ItemAction::Drop(Outcome::Invalid(DiscardReason::Internal));
+                }
+                Annotated(Some(Result::Ok(compat_span)), meta) => {
+                    Annotated(Some(compat_span), meta)
+                }
+                Annotated(None, meta) => Annotated(None, meta),
+            };
+            let payload = match compat_span.to_json() {
+                Ok(payload) => payload,
+                Err(err) => {
+                    relay_log::debug!("failed to serialize compat span: {}", err);
+                    return ItemAction::Drop(Outcome::Invalid(DiscardReason::Internal));
+                }
+            };
 
+            new_item.set_payload(ContentType::CompatSpan, payload);
+        } else {
+            let payload = match annotated_span.to_json() {
+                Ok(payload) => payload,
+                Err(err) => {
+                    relay_log::debug!("failed to serialize span: {}", err);
+                    return ItemAction::Drop(Outcome::Invalid(DiscardReason::Internal));
+                }
+            };
+            new_item.set_payload(ContentType::Json, payload);
+        }
+
+        new_item.set_metrics_extracted(item.metrics_extracted());
         *item = new_item;
 
         ItemAction::Keep
@@ -258,6 +281,15 @@ pub async fn process(
     if let Some(outcome) = sampling_result.into_dropped_outcome() {
         managed_envelope.track_outcome(outcome, DataCategory::SpanIndexed, span_count);
     }
+}
+
+/// Whether or not to convert spans into backward-compatible V2 spans.
+///
+/// This only makes sense when we forward the envelope to Kafka.
+fn produce_compat_spans(config: &Config, global_config: &GlobalConfig, org_id: u64) -> bool {
+    cfg!(feature = "processing")
+        && config.processing_enabled()
+        && utils::is_rolled_out(org_id, global_config.options.span_kafka_v2_sample_rate).is_keep()
 }
 
 fn add_sample_rate(measurements: &mut Annotated<Measurements>, name: &str, value: Option<f64>) {
@@ -296,7 +328,7 @@ pub fn extract_from_event(
     }
 
     if let Some(sample_rate) = global_config.options.span_extraction_sample_rate
-        && sample(sample_rate).is_discard()
+        && utils::sample(sample_rate).is_discard()
     {
         return spans_extracted;
     }